requirements.txt
requests==2.32.3
pandas
google-currency==1.0.10
lxml
bs4
pandas
line-bot-sdk==3.16.2
yfinance==0.2.54

